+++ src/HomeWorkEnum/DayOfWeek.java
@@ -1,0 +1,21 @@
+package HomeWorkEnum;
+
+public enum DayOfWeek {
+    MONDAY(0),
+    WEDNESDAY(10),
+    FRIDAY(15),
+    SUNDAY(20);
+
+    private final int discountPercent;
+
+    DayOfWeek(int discountPercent) {
+        this.discountPercent = discountPercent;
+    }
+
+    public int getDiscountPercent() {
+        return discountPercent;
+    }
+
+
+
+}
